Fix MLX slab API, Metal inhibit kernel, and spike tooling
- Align Python SlabClient read paths with C++ read_slot (no broken like cast).
- Use ushort for bf16 payload clears in MSL (bfloat16_t is invalid on Metal).
- Makefile: maturin/python venv checks and python-clean target.
- Document canonical ~/dev/HiveClaw workflow; pin spike requirements.
- Add swarm_spike with heartbeat, overseer script, VS Code settings.
- Expand .gitignore for venv, .so artifacts, and generated plists.

Made-with: Cursor

+++ scripts/intelligence_spike.py
@@ -132,7 +132,7 @@ def main():
         norm = mx.linalg.norm(
             h_prompt_a.astype(mx.float32), ord=2, axis=-1, keepdims=True
         ).astype(mx.bfloat16)
-        normalized_scent = h_prompt_a / (norm + eps)
+        normalized_scent = (h_prompt_a / (norm + eps)).astype(mx.bfloat16)
 
         write_node = slab_client.write_scent(SLOT_INDEX, normalized_scent)
         mx.eval(write_node)
